requirements.txt
flask==2.3.3
paho-mqtt==2.2.1
pytest==7.4.0
pytest-cov==4.1.0
